pyproject.toml
[project]
name = "cdk-lab-scheduling-delete-all-sagemaker-apps"
version = "0.1.0"
requires-python = ">=3.13.9"
dependencies = [
    "aws-cdk-lib>=2.229.1",
]

[dependency-groups]
dev = [
    "boto3>=1.41.5",
    "ipython>=9.7.0",
]
